+++ src/app/app-routing.module.ts
@@ -5,7 +5,7 @@ import { Roles } from './guards/IsAuthorized/roles';
 
 const routes: Routes = [
   // ? Example of IsAuthorizedGuard usage
-  // { path: 'admin', component: AdminComponent, canActivate: [isAuthorizedGuard], data: { authorize: [Roles.Admin] } }
+  // { path: 'admin', component: AdminComponent, canActivate: [isAuthorizedGuard], data: { authorize: [Roles.Admin], minimunRoleMode: true } },
 ];
 
 @NgModule({